test:api测完

+++ blueprints/group.py
@@ -33,11 +33,20 @@ class Group(Resource):
     def post(self):
         users = request.json.get('users')
         user2_id = request.json.get('user2_id') #user2_id is the owner of the group
+        if user2_id is None or user2_id == "":
+            return jsonify({"message":"user2_id is required", "code":400})
+        if users is None or users == "":
+            return jsonify({"message":"users is required", "code":400})
+        if len(users) < 3:
+            return jsonify({"message":"users is too short", "code":400})
         current_app.logger.info(str(request.remote_addr)+"][User:"+user2_id+" Create Group")
         user2 = UserModel.query.filter(UserModel.id==user2_id).first()
-        name  = request.json.get('name')
         if user2 is None:
-            return jsonify({"message":"user not found"})
+            return jsonify({"message":"user not found", "code":400})
+        name  = request.json.get('name')
+        name = name.replace(" ", "")
+        if name is None or name == "":
+            return jsonify({"message":"name is required", "code":400})
         group = GroupModel(name=name, owner_id=user2_id, create_time=datetime.now())
         db.session.add(group)
         group_member = GroupMemberModel(user_id=user2_id, group_id=group.id)
@@ -46,26 +55,28 @@ class Group(Resource):
         for user in users:
             user = UserModel.query.filter(UserModel.id==user).first()
             if user is None:
-                return jsonify({"message":"user not found"})
+                return jsonify({"message":"user not found", "code":400})
             group_member = GroupMemberModel(user_id=user.id, group_id=group.id)
             db.session.add(group_member)
             db.session.commit()
         current_app.logger.info(str(request.remote_addr)+"][User:"+user2_id+" Create Group Success")
-        return jsonify({"message":"success"})
+        return jsonify({"message":"success", "code":200})
 
     @verifyEmployeeToken
     def get(self):
         current_app.logger.info(str(request.remote_addr)+"][Get Group")
         group_id = request.values.get("group_id")
+        if group_id is None or group_id == "":
+            return jsonify({"message":"group_id is required", "code":400})
         group = GroupModel.query.filter(GroupModel.id==group_id).first()
         if group is None:
-            return jsonify({"message":"group not found"})
+            return jsonify({"message":"group not found", "code":400 })
         members_ = GroupMemberModel.query.filter(GroupMemberModel.group_id==group.id).all()
         members = []
         for member in members_:
             user = UserModel.query.filter(UserModel.id==member.user_id).first()
             members.append({"id":member.user_id, "name":user.username, "avatar": "/api/user/avatar?id=%s" % user.id})
-        return jsonify({"group_id":group.id, "name":group.name, "owner_id":group.owner_id, "members":members})
+        return jsonify({"code":200,"group_id":group.id, "name":group.name, "owner_id":group.owner_id, "members":members})
 
 
 class Message(Resource):
@@ -73,7 +84,13 @@ class Message(Resource):
     def get(self):
         current_app.logger.info(str(request.remote_addr)+"][Get Group Message")
         group_id = request.values.get("group_id")
+        if group_id is None or group_id == "":
+            return jsonify({"message":"group_id is required", "code":400})
+        if not GroupModel.query.filter(GroupModel.id==group_id).first():
+            return jsonify({"message":"group not found", "code":400})
         user_id = decodeToken(request.headers.get("token")).get("id")
+        if user_id is None or user_id == "":
+            return jsonify({"message":"user_id is required", "code":400})
         # session = SessionModel.query.filter(SessionModel.id==session_id)
         # messages = MessageModel.query.filter(MessageModel.session_id==session_id).order_by(MessageModel.id).all()
         messages = GroupMessageModel.query.filter(GroupMessageModel.group_id==group_id).order_by(GroupMessageModel.id).all()
@@ -86,12 +103,18 @@ class Message(Resource):
         if len(his_messages) > 50:
             for i in range(0, len(his_messages)-50):
                 his_messages.pop(i)
-        return jsonify({"messages":his_messages})
+        return jsonify({"messages":his_messages, "code":200})
 
     @verifyEmployeeToken
     def post(self):
         group_id = request.json.get("group_id")
+        if group_id is None or group_id == "":
+            return jsonify({"message":"group_id is required", "code":400})
+        if not GroupModel.query.filter(GroupModel.id==group_id).first():
+            return jsonify({"message":"group not found", "code":400})
         content = request.json.get("content")
+        if content is None:
+            return jsonify({"message":"content is required", "code":400})
         user_id = decodeToken(request.headers.get("token")).get("id")
         dt= datetime.now()
         current_app.logger.info(str(request.remote_addr)+"][User:"+str(user_id)+"Send Message")
@@ -109,21 +132,25 @@ class Message(Resource):
         try:
             db.session.add(message)
             db.session.commit()
-            return "Send message successfully!", 200
+            return jsonify({"message":"success", "code":200})
         except Exception as e:
-            return e, 400
+            return jsonify({"message":e, "code":400})
 
     @verifyEmployeeToken
     def delete(self):
         id = request.values.get("message_id")
+        if id is None or id == "":
+            return jsonify({"message":"message_id is required", "code":400})
         message = GroupMessageModel.query.filter(GroupMessageModel.id==id).first()
+        if message is None:
+            return jsonify({"message":"message not found", "code":400})
         current_app.logger.warning(str(request.remote_addr)+"][Delete Message:"+str(id)+"")
         try:
             db.session.delete(message)
             db.session.commit()
-            return "Delete message successfully!", 200
+            return jsonify({"message":"success", "code":200})
         except Exception as e:
-            return e, 400
+            return jsonify({"message":"delete message failed", "code":400})
 
 
 class Upload(Resource):
@@ -132,9 +159,15 @@ class Upload(Resource):
         current_app.logger.info(str(request.remote_addr)+"][Upload ")
         file = request.files.get('file')
         id = request.headers.get('id')
+        if file is None:
+            return jsonify({"message":"file is required", "code":400})
+        if id is None or id == "":
+            return jsonify({"message":"id is required", "code":400})
         filename=file.filename
         filetype=filename.split(".")[-1]
         message = GroupMessageModel.query.filter(GroupMessageModel.id==id).first()
+        if message is None:
+            return jsonify({"message":"message not found", "code":400})
         if filetype == "png" or filetype == "jpg" or filetype == "jpeg":
             file.save("asset/group/files/"+id+"."+filetype)
             type = "image"
@@ -149,10 +182,9 @@ class Upload(Resource):
         message.url=url
         try:
             db.session.commit()
-            return "Successfully!", 200
+            return jsonify({"message":"success", "code":200})
         except Exception as e:
-            return e, 400
-
+            return jsonify({"message":"upload failed", "code":400})
     def get(self):
         filename = request.args.get('filename')
         img_local_path = "./asset/group/files/" + filename
@@ -163,15 +195,24 @@ class Upload(Resource):
             img_f.close()
             return res
         except:
-            return "File not found!", 404
+            return jsonify({"message":"file not found", "code":400})
 
 class updateFileContent(Resource):
     @verifyEmployeeToken
     def post(self):
         current_app.logger.info(str(request.remote_addr)+"][Upload Group file content")
         group_id = request.json.get("group_id")
+        if group_id is None or group_id == "":
+            return jsonify({"message":"group_id is required", "code":400})
+        if not GroupModel.query.filter(GroupModel.id==group_id).first():
+            return jsonify({"message":"group not found", "code":400})
         content = request.json.get("content")
-        user_id = decodeToken(request.headers.get("token")).get("id")
+        if content is None:
+            return jsonify({"message":"content is required", "code":400})
+        user_id = decodeToken(request.headers.get("token"))
+        if user_id is None:
+            return jsonify({"message":"token is invalid", "code":400})
+        user_id = user_id.get("id")
         dt= datetime.now()
         year=dt.year
         month=dt.month
@@ -186,9 +227,9 @@ class updateFileContent(Resource):
         try:
             db.session.add(message)
             db.session.commit()
-            return jsonify({"id": message.id})
+            return jsonify({"id": message.id, "message":"success", "code":200})
         except Exception as e:
-            return e, 400
+            return jsonify({"message":e, "code":400})
     @verifyEmployeeToken
     def get(self):
         filename = request.args.get('filename')
@@ -196,7 +237,7 @@ class updateFileContent(Resource):
         try:
             return send_file(img_local_path, as_attachment=True, attachment_filename=filename)
         except:
-            return "File not found!", 404
+            return jsonify({"message":"file not found", "code":400})
 
 api.add_resource(Group, "/group")
 api.add_resource(Message, "/message")
